requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llvmlite==0.39.1
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
narwhals==2.19.0
networkx==3.4.2
numba==0.56.4
numpy==1.23.5
openai-whisper==20250625
packaging==26.0
pandas==2.3.3
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sounddevice==0.5.5
soundfile==0.13.1
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
torch==2.2.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
